% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_db.R
\name{call_db}
\alias{call_db}
\title{call_db}
\usage{
call_db(tableName, hostname = "localhost", username, pass, database,
  path = "temp/")
}
\arguments{
\item{tableName}{The table name.}

\item{hostname}{The host name of the database. Default set to localhost}

\item{username}{The username to access the database.}

\item{pass}{The password to access the database}

\item{database}{The name of the database.}

\item{path}{File path. Defaults set as "temp"}
}
\value{
Returns a dataframe
}
\description{
Deploy a table in an R directory, as an RDA file.
}
