% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catdeploy.R
\name{catdeploy}
\alias{catdeploy}
\title{catdeploy}
\usage{
catdeploy(bin, hostname = "localhost", username = "tcatdbuser", pass,
  database = "twittercapture", extension = "tsv", path = "./")
}
\arguments{
\item{bin}{The bin name}

\item{hostname}{The host of the database. Default set to localhost}

\item{username}{The name of the database. Default set to tcatdbuser}

\item{pass}{Password of the database}

\item{database}{The name of the database. Default set to twittercapture}

\item{extension}{What kind of file to you wand to work with ? Default set to a tsv format}

\item{path}{Path of the data file. Defaults set to current directory}
}
\value{
Return csv/tsv file, with an UTF-8 encoding
}
\description{
Deploy a TCAT database as a csv/tsv file, with an UTF-8 encoding
}
