% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcatdeploy.R
\name{tcatdeploy}
\alias{tcatdeploy}
\title{tcatdeploy}
\usage{
tcatdeploy(bin, hostname = "localhost", username = "tcatdbuser", pass,
  database = "twittercapture", deploy = TRUE, extension = "rda",
  path = "temp/")
}
\arguments{
\item{bin}{The bin name}

\item{hostname}{The host name of the database. Default set to localhost}

\item{username}{The username to access the database. Default set to tcatdbuser}

\item{pass}{The password to access the database}

\item{database}{The name of the database. Default set to twittercapture}

\item{deploy}{Save the datas as a file. Default set to TRUE}

\item{extension}{The chosen extension for the deployment of the database. Possible choices : "csv", "tsv", "rda". Default set to a "rda" format}

\item{path}{File path. Defaults set to current directory}
}
\value{
Returns a csv / tsv file, with a UTF-8 encoding
}
\description{
Deploy a DMI-TCAT database in an R directory as a csv / tsv file with UTF-8 encoding.
}
