% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcattempo.R
\name{tcattempo}
\alias{tcattempo}
\title{tcattempo}
\usage{
tcattempo(bin, hostname = "localhost", username = "tcatdbuser", pass,
  database = "twittercapture", extension = "tsv", path = "./", startday,
  endday, starttime = "00:00:00", endtime = "23:59:59")
}
\arguments{
\item{bin}{The bin name}

\item{hostname}{The host of the database. Default set to localhost}

\item{username}{The name of the database. Default set to tcatdbuser}

\item{pass}{Password of the database}

\item{database}{The name of the database. Default set to twittercapture}

\item{extension}{What kind of file to you wand to work with ? Default set to a tsv format}

\item{path}{Path of the data file. Defaults set to current directory}

\item{startday}{The first day of the period you want to select}

\item{endday}{The last day of the period you want to select}

\item{starttime}{The start time of the period you want to select. Default set to 00:00:00}

\item{endtime}{The end time of the period you want to select. Default set to 23:59:59}
}
\value{
Return csv/tsv file, with an UTF-8 encoding
}
\description{
Deploy a TCAT database as a csv/tsv file, with an UTF-8 encoding and a selection of a time period
}
