% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcattempo.R
\name{tcattempo}
\alias{tcattempo}
\title{tcattempo}
\usage{
tcattempo(bin, hostname = "localhost", username = "tcatdbuser", pass,
  database = "twittercapture", deploy = TRUE, extension = "tsv",
  path = "./", startday, endday = Sys.Date(), starttime = "00:00:00",
  endtime = "23:59:59")
}
\arguments{
\item{bin}{The bin name}

\item{hostname}{The host name of the database. Default set to localhost}

\item{username}{The username to access the database. Default set to tcatdbuser}

\item{pass}{The password to access the database}

\item{database}{The name of the database. Default set to twittercapture}

\item{deploy}{Save the datas as a file. Default set to TRUE}

\item{extension}{The chosen extension for the deployment of the database. Possible choices : "csv", "tsv". Default set to a "tsv" format}

\item{path}{File path. Defaults set to current directory}

\item{startday}{The first day of the period you want to select}

\item{endday}{The last day of the period you want to select}

\item{starttime}{The time of the first day of the period you want to select. Default set to 00:00:00}

\item{endtime}{The time of the last day of the period you want to select. Default set to 23:59:59}
}
\value{
Returns a csv / tsv file, with a UTF-8 encoding
}
\description{
Deploy a DMI-TCAT database in an R directory, as a csv / tsv file, with UTF-8 encoding and a period selection.
}
